Introducing Assembler, can craft its own bills.
= Rationale
- Building_Assembler
Inheriting from Building_WorkTable we are able to self craft the
recipes in the ITab_Bills with no downsides.

- Building_Crafter:
By moving the Crafter under the Assembler, inheritance is kept sane. The
defs don't change.

= Usage
Requires ITab_Bills to function.

<ThingDef Name="SALAssemblerBase" ParentName="SALBase" Abstract="True">
    <thingClass>ProjectSAL.Building_Assembler</thingClass>
    <hasInteractionCell>false</hasInteractionCell>
    <inspectorTabs>
        <li>ITab_Bills</li>
    </inspectorTabs>
</ThingDef>

= Issues
Buildings that are bigger than 1x1 have more output positions. I changed
the rotOutput variable, will cause Save compatibility issues on update.
