+++ Source/Building_Assembler_LogicHelpers.cs
@@ -7,7 +7,7 @@ using Verse;
 
 namespace ProjectSAL
 {
-    public partial class Building_Crafter
+    public partial class Building_Assembler
     {
         public static Thing CalculateDominantIngredient(RecipeDef currentRecipe, List<Thing> thingRecord)
         {
