+++ Source/Source.csproj
@@ -20,6 +20,7 @@
     <NoStdLib>False</NoStdLib>
     <TreatWarningsAsErrors>False</TreatWarningsAsErrors>
     <IntermediateOutputPath>obj\$(Configuration)\</IntermediateOutputPath>
+    <UseMSBuildEngine>false</UseMSBuildEngine>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <DebugSymbols>false</DebugSymbols>
@@ -29,6 +30,7 @@
     <DefineConstants>DEBUG</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
+    <PlatformTarget>anycpu</PlatformTarget>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
     <DebugType>none</DebugType>
@@ -49,35 +51,37 @@
     <GenerateSerializationAssemblies>Auto</GenerateSerializationAssemblies>
   </PropertyGroup>
   <ItemGroup>
-    <Reference Include="Assembly-CSharp">
-      <HintPath>..\..\..\RimWorldWin_Data\Managed\Assembly-CSharp.dll</HintPath>
-      <Private>False</Private>
-    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
     <Reference Include="System.Data" />
     <Reference Include="System.Xml" />
+    <Reference Include="Assembly-CSharp">
+      <HintPath>..\..\..\RimWorld1557Linux_Data\Managed\Assembly-CSharp.dll</HintPath>
+      <Private>False</Private>
+    </Reference>
     <Reference Include="UnityEngine">
-      <HintPath>..\..\..\RimWorldWin_Data\Managed\UnityEngine.dll</HintPath>
+      <HintPath>..\..\..\RimWorld1557Linux_Data\Managed\UnityEngine.dll</HintPath>
       <Private>False</Private>
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Building_Crafter_AcceptingItems.cs" />
-    <Compile Include="Building_Crafter.cs" />
-    <Compile Include="Building_Crafter_LogicHelpers.cs" />
-    <Compile Include="Building_Crafter_MiscOverride.cs" />
-    <Compile Include="Building_Crafter_Properties.cs" />
-    <Compile Include="Building_Crafter_Ticking.cs" />
     <Compile Include="Building_SmartHopper.cs" />
-    <Compile Include="Building_Crafter_Pawn.cs" />
     <Compile Include="Dialogs.cs" />
     <Compile Include="Defs.cs" />
     <Compile Include="ModSettings.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ProjectSAL_Utilities.cs" />
+    <Compile Include="Building_Assembler.cs" />
+    <Compile Include="Building_Assembler_AcceptingItems.cs" />
+    <Compile Include="Building_Assembler_LogicHelpers.cs" />
+    <Compile Include="Building_Assembler_MiscOverride.cs" />
+    <Compile Include="Building_Assembler_Pawn.cs" />
+    <Compile Include="Building_Assembler_Properties.cs" />
+    <Compile Include="Building_Assembler_Ticking.cs" />
+    <Compile Include="Building_Crafter.cs" />
+    <Compile Include="ResourceBank.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
@@ -87,4 +91,11 @@
   <Target Name="AfterBuild">
   </Target>
   -->
+  <ProjectExtensions>
+    <MonoDevelop>
+      <Properties>
+        <Deployment.LinuxDeployData generatePcFile="False" />
+      </Properties>
+    </MonoDevelop>
+  </ProjectExtensions>
 </Project>
